requirements.txt
fastapi[all]
sqlalchemy
alembic
pymysql
python-jose
passlib[bcrypt]
cryptography
python-dotenv